+++ tencentcloud/lighthouse/v20200324/errors.go
@@ -44,6 +44,12 @@ const (
 	// 查询镜像失败，请稍后再试。
 	FAILEDOPERATION_DESCRIBEBLUEPRINTSFAILED = "FailedOperation.DescribeBlueprintsFailed"
 
+	// 查询套餐折扣失败
+	FAILEDOPERATION_DESCRIBEBUNDLEDISCOUNTFAILED = "FailedOperation.DescribeBundleDiscountFailed"
+
+	// 查询套餐失败
+	FAILEDOPERATION_DESCRIBEBUNDLESFAILED = "FailedOperation.DescribeBundlesFailed"
+
 	// 查询实例状态错误。
 	FAILEDOPERATION_DESCRIBEINSTANCESTATUS = "FailedOperation.DescribeInstanceStatus"