+++ tencentcloud/tcss/v20201101/models.go
@@ -10844,6 +10844,9 @@ type DescribeAssetImageRegistryDetailResponseParams struct {
 	// 注意：此字段可能返回 null，表示取不到有效值。
 	SensitiveInfoCnt *uint64 `json:"SensitiveInfoCnt,omitnil" name:"SensitiveInfoCnt"`
 
+	// Id
+	Id *uint64 `json:"Id,omitnil" name:"Id"`
+
 	// 唯一请求 ID，每次请求都会返回。定位问题时需要提供该次请求的 RequestId。
 	RequestId *string `json:"RequestId,omitnil" name:"RequestId"`
 }