+++ tencentcloud/tiw/v20190919/client.go
@@ -1055,7 +1055,7 @@ func NewDescribePPTCheckResponse() (response *DescribePPTCheckResponse) {
 }
 
 // DescribePPTCheck
-// 查询PPT检测任务的执行进度或结果
+// 查询PPT检测任务的执行进度或结果，支持查询最近半年内的任务结果
 //
 // 可能返回的错误码:
 //  AUTHFAILURE = "AuthFailure"
@@ -1078,7 +1078,7 @@ func (c *Client) DescribePPTCheck(request *DescribePPTCheckRequest) (response *D
 }
 
 // DescribePPTCheck
-// 查询PPT检测任务的执行进度或结果
+// 查询PPT检测任务的执行进度或结果，支持查询最近半年内的任务结果
 //
 // 可能返回的错误码:
 //  AUTHFAILURE = "AuthFailure"