+++ tencentcloud/vpc/v20170312/models.go
@@ -25088,7 +25088,9 @@ type Route struct {
 	GatewayType *string `json:"GatewayType,omitnil" name:"GatewayType"`
 
 	// 下一跳地址，这里只需要指定不同下一跳类型的网关ID，系统会自动匹配到下一跳地址。
-	// 特殊说明：GatewayType为NORMAL_CVM时，GatewayId填写实例的内网IP。
+	// 特殊说明：
+	// GatewayType为NORMAL_CVM时，GatewayId填写实例的内网IP。
+	// GatewayType为EIP时，GatewayId填写0。
 	GatewayId *string `json:"GatewayId,omitnil" name:"GatewayId"`
 
 	// 路由策略ID。IPv4路由策略ID是有意义的值，IPv6路由策略是无意义的值0。后续建议完全使用字符串唯一ID `RouteItemId`操作路由策略。
@@ -25112,6 +25114,7 @@ type Route struct {
 	RouteTableId *string `json:"RouteTableId,omitnil" name:"RouteTableId"`
 
 	// 目的IPv6网段，取值不能在私有网络网段内，例如：2402:4e00:1000:810b::/64。
+	// 注意：此字段可能返回 null，表示取不到有效值。
 	DestinationIpv6CidrBlock *string `json:"DestinationIpv6CidrBlock,omitnil" name:"DestinationIpv6CidrBlock"`
 
 	// 路由唯一策略ID。